pyproject.toml
[project]
name = "stock-analysis-report"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.0",
    "python-dotenv>=1.2.1",
    "yfinance>=0.2.66",
]
